Rename source file import statements to match team name
