+++ src/main/java/org/commitment_issues/Start.java
@@ -1,8 +1,8 @@
-package org.yourteamname;
+package org.commitment_issues;
 
 import java.util.List;
 import java.util.Vector;
-import org.yourteamname.agents.DummyAgent;
+import org.commitment_issues.agents.DummyAgent;
 
 public class Start {
     public static void main(String[] args) {
